+++ src/components/loginForm.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import Form from "./common/form";
 import Joi from "joi-browser";
-import * as authService from "../services/authService";
+import auth from "../services/authService";
 
 class LoginForm extends Form {
   state = {
@@ -20,11 +20,8 @@ class LoginForm extends Form {
   doSubmit = async () => {
     try {
       const { data } = this.state;
-      const { data: jwt } = await authService.login(
-        data.username,
-        data.password
-      );
-      localStorage.setItem("token", jwt);
+      await auth.login(data.username, data.password);
+
       // redirect to home page
       window.location = "/";
     } catch (ex) {
