+++ src/components/logout.jsx
@@ -1,8 +1,9 @@
 import React, { Component } from "react";
+import auth from "../services/authService";
 
 class Logout extends Component {
   componentDidMount() {
-    localStorage.removeItem("token");
+    auth.logout();
     window.location = "/";
   }
   render() {
